apply asyncfeature on whisper ios
