+++ whisper/ZeticMLangeWhisper-iOS/ZeticMLange.framework/_CodeSignature/CodeResources
@@ -6,31 +6,31 @@
 	<dict>
 		<key>Info.plist</key>
 		<data>
-		Wt0+qzeAeZRu5aKTeryGGIAU+rU=
+		l9sxFkguH+6N2IhPSbwXrgeuFro=
 		</data>
 		<key>Modules/ZeticMLange.swiftmodule/Project/arm64-apple-ios.swiftsourceinfo</key>
 		<data>
-		H/2a3SjdUO8kTXnwr5zb4EYu5g0=
+		2BBPsdyARoMIIRjZAQianN7spHI=
 		</data>
 		<key>Modules/ZeticMLange.swiftmodule/arm64-apple-ios.abi.json</key>
 		<data>
-		BioJ1fUdk4zV+XWFIhEdEb1UOgI=
+		42MF3M9bhzKPN/3xJkX0EMnkCAE=
 		</data>
 		<key>Modules/ZeticMLange.swiftmodule/arm64-apple-ios.private.swiftinterface</key>
 		<data>
-		9J8qtLH7wqzpbyRxskAWuEUl50M=
+		PNtZ0Ch2IVHfl1wV+lvr8X8zGQM=
 		</data>
 		<key>Modules/ZeticMLange.swiftmodule/arm64-apple-ios.swiftdoc</key>
 		<data>
-		JOZrhIIJGTiDv4EMTDeDOUmnWyY=
+		48LzpYNIYTiZbgdXeQaHNYK0aXs=
 		</data>
 		<key>Modules/ZeticMLange.swiftmodule/arm64-apple-ios.swiftinterface</key>
 		<data>
-		9J8qtLH7wqzpbyRxskAWuEUl50M=
+		PNtZ0Ch2IVHfl1wV+lvr8X8zGQM=
 		</data>
 		<key>Modules/ZeticMLange.swiftmodule/arm64-apple-ios.swiftmodule</key>
 		<data>
-		uQJC+1vYNeAJXQVL18pBhONw7i0=
+		V9w0JZdwB0/aQNlGQH8wC9NCzXA=
 		</data>
 	</dict>
 	<key>files2</key>
@@ -39,66 +39,66 @@
 		<dict>
 			<key>hash</key>
 			<data>
-			H/2a3SjdUO8kTXnwr5zb4EYu5g0=
+			2BBPsdyARoMIIRjZAQianN7spHI=
 			</data>
 			<key>hash2</key>
 			<data>
-			vAlqVzKgl7cIkaJOxbjImtIqhgG6kklXGh/mx2+OkJg=
+			3CY6Za8hZQyCiNMdibJwFWiNl0nI6uDlLEadFqphnFQ=
 			</data>
 		</dict>
 		<key>Modules/ZeticMLange.swiftmodule/arm64-apple-ios.abi.json</key>
 		<dict>
 			<key>hash</key>
 			<data>
-			BioJ1fUdk4zV+XWFIhEdEb1UOgI=
+			42MF3M9bhzKPN/3xJkX0EMnkCAE=
 			</data>
 			<key>hash2</key>
 			<data>
-			76Djr6+vM67oOY/VeOmnTLSi2VseA6UeKd1WmIYXRes=
+			NhfyCiQCxMBsD1LUx+3cgDE5l5A3rGp4EHS1QQqloeI=
 			</data>
 		</dict>
 		<key>Modules/ZeticMLange.swiftmodule/arm64-apple-ios.private.swiftinterface</key>
 		<dict>
 			<key>hash</key>
 			<data>
-			9J8qtLH7wqzpbyRxskAWuEUl50M=
+			PNtZ0Ch2IVHfl1wV+lvr8X8zGQM=
 			</data>
 			<key>hash2</key>
 			<data>
-			jpp8q4l6mvX35KzQ4OouysmCD7ca/A4SHbx22TljHUY=
+			eUJF0F4YispGrblZ0O8QIJ+KMj5PoL1IHsvTaA7rU/w=
 			</data>
 		</dict>
 		<key>Modules/ZeticMLange.swiftmodule/arm64-apple-ios.swiftdoc</key>
 		<dict>
 			<key>hash</key>
 			<data>
-			JOZrhIIJGTiDv4EMTDeDOUmnWyY=
+			48LzpYNIYTiZbgdXeQaHNYK0aXs=
 			</data>
 			<key>hash2</key>
 			<data>
-			kpC4WBy/zUfWGLYYqAM4g4TZGcGT5olI39wXtYXktEI=
+			nnKIs83YS50RLRDOhnU3MdBSUz5tfKr7SY0ANK2Z90c=
 			</data>
 		</dict>
 		<key>Modules/ZeticMLange.swiftmodule/arm64-apple-ios.swiftinterface</key>
 		<dict>
 			<key>hash</key>
 			<data>
-			9J8qtLH7wqzpbyRxskAWuEUl50M=
+			PNtZ0Ch2IVHfl1wV+lvr8X8zGQM=
 			</data>
 			<key>hash2</key>
 			<data>
-			jpp8q4l6mvX35KzQ4OouysmCD7ca/A4SHbx22TljHUY=
+			eUJF0F4YispGrblZ0O8QIJ+KMj5PoL1IHsvTaA7rU/w=
 			</data>
 		</dict>
 		<key>Modules/ZeticMLange.swiftmodule/arm64-apple-ios.swiftmodule</key>
 		<dict>
 			<key>hash</key>
 			<data>
-			uQJC+1vYNeAJXQVL18pBhONw7i0=
+			V9w0JZdwB0/aQNlGQH8wC9NCzXA=
 			</data>
 			<key>hash2</key>
 			<data>
-			y0oENIAps3ZlYq9s7wp7jMhu6qUfLKI/4ODnkh0hTu4=
+			DGndV2QAFryCzaH7TpRk4nezLhiJyhxfvFnjFihRm8s=
 			</data>
 		</dict>
 	</dict>
