+++ whisper/ZeticMLangeWhisper-iOS/ZeticMLangeWhisper-iOS/ZeticMLangeWhisper_iOSApp.swift
@@ -4,7 +4,7 @@ import SwiftUI
 struct ZeticMLangeWhisper_iOSApp: App {
     var body: some Scene {
         WindowGroup {
-            ContentView()
+            WhisperView()
         }
     }
 }
